% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pretty_bytes}
\alias{pretty_bytes}
\title{Bytes in a human readable string}
\usage{
pretty_bytes(bytes)
}
\arguments{
\item{bytes}{Numeric vector.}
}
\value{
Character vector, the formatted sizes.
}
\description{
Bytes in a human readable string
}

