% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pretty_dt}
\alias{pretty_dt}
\title{Pretty formatting of time intervals (difftime objects)}
\usage{
pretty_dt(dt, compact = FALSE)
}
\arguments{
\item{dt}{A \code{difftime} object, a vector of time
differences.}

\item{compact}{If true, then only the first non-zero
unit is used. See examples below.}
}
\value{
Character vector of formatted time intervals.
}
\description{
Pretty formatting of time intervals (difftime objects)
}
\examples{
pretty_dt(as.difftime(1000, units = "secs"))
pretty_dt(as.difftime(0, units = "secs"))
}
\seealso{
Other time: \code{\link{pretty_ms}};
  \code{\link{pretty_sec}}
}

