% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pretty_ms}
\alias{pretty_ms}
\title{Pretty formatting of milliseconds}
\usage{
pretty_ms(ms, compact = FALSE)
}
\arguments{
\item{ms}{Numeric vector of milliseconds}

\item{compact}{If true, then only the first non-zero
unit is used. See examples below.}
}
\value{
Character vector of formatted time intervals.
}
\description{
Pretty formatting of milliseconds
}
\examples{
pretty_ms(c(1337, 13370, 133700, 1337000, 1337000000))

pretty_ms(c(1337, 13370, 133700, 1337000, 1337000000),
          compact = TRUE)
}
\seealso{
Other time: \code{\link{pretty_dt}};
  \code{\link{pretty_sec}}
}

