% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pretty_sec}
\alias{pretty_sec}
\title{Pretty formatting of seconds}
\usage{
pretty_sec(sec, compact = FALSE)
}
\arguments{
\item{sec}{Numeric vector of seconds.}

\item{compact}{If true, then only the first non-zero
unit is used. See examples below.}
}
\value{
Character vector of formatted time intervals.
}
\description{
Pretty formatting of seconds
}
\examples{
pretty_sec(c(1337, 13370, 133700, 1337000, 13370000))

pretty_sec(c(1337, 13370, 133700, 1337000, 13370000),
           compact = TRUE)
}
\seealso{
Other time: \code{\link{pretty_dt}};
  \code{\link{pretty_ms}}
}

