% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-package.R
\docType{package}
\name{progress-package}
\alias{progress}
\alias{progress-package}
\title{progress: Terminal Progress Bars}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Configurable Progress bars, they may include percentage, elapsed time, and/or the estimated completion time. They work in terminals, in 'Emacs' 'ESS', 'RStudio', 'Windows' 'Rgui' and the 'macOS' 'R.app'. The package also provides a 'C++' 'API', that works with or without 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/progress#readme}
  \item \url{http://r-lib.github.io/progress/}
  \item Report bugs at \url{https://github.com/r-lib/progress/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Rich FitzJohn
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
