% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progress.r
\docType{data}
\name{progress_bar}
\alias{progress_bar}
\title{Progress bar on the terminal}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fd254921d08> 
 - attr(*, "name")= chr "progress_bar_generator"
}}
\usage{
progress_bar
}
\description{
Progress bar on the terminal
}
\examples{
\dontrun{
pb <- progress_bar$new(total = 100)
for (i in 1:100) {
  pb$tick()
  Sys.sleep(1 / 100)
}

pb <- progress_bar$new(
  format = "  downloading [:bar] :percent eta: :eta",
  total = 100, clear = FALSE, width= 60)
for (i in 1:100) {
  pb$tick()
  Sys.sleep(1 / 100)
}

pb <- progress_bar$new(
  format = "  downloading [:bar] :percent in :elapsed",
  total = 100, clear = FALSE, width= 60)
for (i in 1:100) {
  pb$tick()
  Sys.sleep(1 / 100)
}
}
}
\keyword{datasets}

