% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate}
\alias{replicate}
\title{Same as \code{\link[base:lapply]{base::replicate()}} but with a progress bar.}
\usage{
replicate(
  n,
  expr,
  simplify = "array",
  format = "Replicating : [:bar] :percent in :elapsed",
  ...
)
}
\arguments{
\item{n}{Integer: the number of replications.}

\item{expr}{The expression (a language object, usually a call) to
evaluate repeatedly.}

\item{simplify}{Passed to \code{\link[base:lapply]{base::sapply()}}, see the docs there.}

\item{format}{Progress bar format.}

\item{...}{Extra arguments are passed to \code{progress_bar$new()}.}
}
\description{
Same as \code{\link[base:lapply]{base::replicate()}} but with a progress bar.
}
\examples{
replicate(50000, rnorm(1) > 1.65) |>
    (\(x) list(mean = mean(x), sd = sd(x)))()
}
