% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_pb}
\alias{replicate_pb}
\title{Same as \code{\link{base::replicate}} but with a progress bar.}
\usage{
replicate_pb(
  n,
  expr,
  simplify = "array",
  format = "Replicating : [:bar] :percent in :elapsed",
  ...
)
}
\description{
Same as \code{\link{base::replicate}} but with a progress bar.
}
\examples{
library(dplyr)
replicate_pb(50000, {rnorm(1)>1.65}) \%>\% {paste(mean(.), " - ", sd(.))}
}
\seealso{
\code{\link{s2a}}
}
