% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{signs}
\alias{signs}
\title{Various handy symbols to use in a command line UI}
\format{A named list, see \code{names(signs)} for all sign names.}
\usage{
signs
}
\description{
Various handy symbols to use in a command line UI
}
\details{
On Windows they have a fallback to less fancy symbols.
}
\examples{
cat(signs$tick, " SUCCESS\\n", signs$cross, " FAILURE\\n", sep="")

## All signs
cat(paste(format(names(signs), width=20),
  unlist(signs)), sep = "\\n")
}
\keyword{datasets}

