% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spark_line}
\alias{spark_line}
\title{Spark line of a numeric vector on the terminal}
\usage{
spark_line(data)
}
\arguments{
\item{data}{The data to visualize. It can be a numeric
vector, or anything that can be cut into intervals
with \code{cut}. Infinite values in numeric data are ignored,
and a black character is plotted instead of them.}
}
\value{
Character scalar containing the spark line.
}
\description{
Spark line of a numeric vector on the terminal
}
\examples{
## Annoal number of Lynx trappings
cat(spark_line(lynx[1:getOption("width")]), "\\n")

## Luteinizing Hormone in Blood Samples,
## in blue, if the terminal supports it
cat(crayon::blue(spark_line(lh)), "\\n")
}

